requirements.txt
# 基础依赖
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.0
pandas
requests[socks]>=2.31.0

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangChain CAPL生成器依赖
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
